Move deposit-fee-calculator under wallets
