+++ src/app/wallets/update-on-chain-receipt.ts
@@ -4,7 +4,8 @@ import { PriceService } from "@services/price"
 import { NotificationsService } from "@services/notifications"
 import { LedgerService } from "@services/ledger"
 import { OnChainError, TxDecoder } from "@domain/bitcoin/onchain"
-import { toLiabilitiesAccountId, DepositFeeCalculator } from "@domain/ledger"
+import { toLiabilitiesAccountId } from "@domain/ledger"
+import { DepositFeeCalculator } from "@domain/wallets"
 import { LockService } from "@services/lock"
 import { ONCHAIN_LOOK_BACK, ONCHAIN_MIN_CONFIRMATIONS, BTC_NETWORK } from "@config/app"
